+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # TBD
 ### `nstd`
+- Added `extern "C"` to `NSTDAPI` (removes `NSTDCPP[START|END]`).
+### `nstd.core`
+- Removed `nstd_core_slice[_mut]_compare`.
+
+# 0.1.2
+### `nstd`
 - Removed `Const` postfix from `NSTDAnyConst`.
 ### `nstd.core`
+- Fixed `NSTDSlice[Mut]::as_slice[_mut]` returning a slice of incorrect length.
 - Renamed `fty` functions.
 - Added `nstd_core_str[_mut]_byte_len`.
 - Removed `Const` postfix from `NSTD[Ptr|Slice|Str|CStr]Const`.